// frontend/src/services/generated/model/playerResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface PlayerResponse {
  /** Hráčova unikátní přezdívka */
  nickname?: string;
  /** Unikátní ID hráče */
  playerId?: number;
  /** Hráčovo aktuální skóre */
  score?: number;
}